homepage logout

+++ controllers/api/tickets.js
@@ -37,7 +37,13 @@ async function index(req, res) {
 
 async function deleteTicket(req, res) {
   try {
-    let ticket = await Ticket.findByIdAndDelete(req.params.id);
+    console.log(req.user._id, "<--user id");
+    let ticket = await Ticket.findById(req.params.id);
+
+    console.log(ticket.reporter, "<--reporter id");
+    if (req.user._id == ticket.reporter) {
+      ticket = await Ticket.findByIdAndDelete(req.params.id);
+    }
     res.status(200).json("Successfully deleted");
   } catch (err) {
     res.status(500).json(err);
